+++ Adafruit_GFX.h
@@ -109,7 +109,7 @@ class Adafruit_GFX_Button {
  public:
   Adafruit_GFX_Button(void);
   void initButton(Adafruit_GFX *gfx, int16_t x, int16_t y,
-   uint8_t w, uint8_t h, uint16_t outline, uint16_t fill,
+   uint16_t w, uint16_t h, uint16_t outline, uint16_t fill,
    uint16_t textcolor, char *label, uint8_t textsize);
   void drawButton(boolean inverted = false);
   boolean contains(int16_t x, int16_t y);
